Add XLateEscapeChars
