+++ VS/VS_2019/MlbDev2/Utility/Utility.vcxproj
@@ -102,6 +102,7 @@
     <ClInclude Include="..\..\..\..\include\Utility\ValueToStringRadix.hpp" />
     <ClInclude Include="..\..\..\..\include\Utility\ValueTraits.hpp" />
     <ClInclude Include="..\..\..\..\include\Utility\VersionNumber.hpp" />
+    <ClInclude Include="..\..\..\..\include\Utility\XLateEscapeChars.hpp" />
     <ClInclude Include="..\..\..\..\include\Utility\XLateToHuman.hpp" />
   </ItemGroup>
   <ItemGroup>
@@ -166,6 +167,7 @@
     <ClCompile Include="..\..\..\..\Utility\UniqueId.cpp" />
     <ClCompile Include="..\..\..\..\Utility\UserName.cpp" />
     <ClCompile Include="..\..\..\..\Utility\VersionNumber.cpp" />
+    <ClCompile Include="..\..\..\..\Utility\XLateEscapeChars.cpp" />
   </ItemGroup>
   <PropertyGroup Label="Globals">
     <VCProjectVersion>16.0</VCProjectVersion>
