+++ VS/VS_2019/MlbDev2/Utility/Utility.vcxproj.filters
@@ -267,6 +267,9 @@
     <ClInclude Include="..\..\..\..\include\Utility\ParseCsv.hpp">
       <Filter>Header Files</Filter>
     </ClInclude>
+    <ClInclude Include="..\..\..\..\include\Utility\XLateEscapeChars.hpp">
+      <Filter>Header Files</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="..\..\..\..\Utility\nstrcat.cpp">
@@ -450,6 +453,9 @@
       <Filter>Source Files</Filter>
     </ClCompile>
     <ClCompile Include="..\..\..\..\Utility\ParseCsv.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="..\..\..\..\Utility\XLateEscapeChars.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
   </ItemGroup>
